+++ src/errors/error.js
@@ -1,9 +1,0 @@
-export default class ErrorClass {
-  constructor(message) {
-    Error.captureStackTrace(this, this.constructor);
-
-    Object.defineProperty(this, 'message', {
-      value: message,
-    });
-  }
-}