+++ src/exceptions/API/gateway_timeout_error.js
@@ -1,0 +1,3 @@
+import EasyPostException from '../easypost_exception';
+
+export default class GatewayTimeoutError extends EasyPostException {}