+++ src/exceptions/API/method_not_allowed_error.js
@@ -1,0 +1,3 @@
+import EasyPostException from '../easypost_exception';
+
+export default class MethodNotAllowedError extends EasyPostException {}