+++ src/exceptions/API/rate_limit_error.js
@@ -1,0 +1,3 @@
+import EasyPostException from '../easypost_exception';
+
+export default class RateLimitError extends EasyPostException {}