+++ src/exceptions/API/unauthorized_error.js
@@ -1,0 +1,3 @@
+import EasyPostException from '../easypost_exception';
+
+export default class UnauthorizedError extends EasyPostException {}