+++ test/services/easypost.test.js
@@ -1,13 +1,14 @@
 import { expect } from 'chai';
 
 import EasyPost from '../../src/easypost';
+import MissingParameterError from '../../src/exceptions/General/missing_parameter_error';
 
 /* eslint-disable func-names */
 describe('EasyPost', function () {
   it('throws an error when no API key is provided', async function () {
     expect(() => new EasyPost()).to.throw(
-      Error,
-      'No API key supplied. Pass in an API key as the first argument.',
+      MissingParameterError,
+      'Missing required parameter: API Key.',
     );
   });
 });