+++ CHANGELOG.md
@@ -24,6 +24,9 @@
 - Changes the `primaryOrSecondary` parameter name to `priority` in billing functions to match the API
 - Empty response functions now return nothing (deleting records, funding a wallet, etc)
 - Explicit `type` parameter for retrieving all reports removed, include `type` in generic `params` dictionary instead.
+- Improves error exception handling
+  - Specific error types for each category of error
+  - API error message may be an array rather than a string. Arrays will be concatenated (by comma) and returned as a string.
 
 ## v5.10.0 (2023-01-18)